Tenant ports not unplugged during instance destroy
When an instance is destroyed, ironic unbinds its ports by unsetting the
host ID and binding:profile fields. The networking-generic-switch driver
only deconfigures switch ports in the delete operation, at which point
the binding information required to do this is lost.  The fix is to
handle unbinding during port updates and deconfigure the switch ports at
that point.