Workaround switch SSH connection rate limiting
On many switches, there are limits to the number of concurrent SSH
sessions and/or the rate of connection attempts. When performing many
operations in a short space of time it is possible to hit these limits.

This change attempts to work around these issues by adding a retry
mechanism to the establishment of SSH connections. There are clearly
limitations to this solution, including the possibility of operations
taking a very long time when SSH connections are rate limited and in
severe cases, hitting the retry limit and failing.

Change-Id: I5aa682e06a0125b4c2ad3e454f8ffe23222259ea
Related-Bug: #1674324