Support configuration of trunk ports
Trunk ports are necessary to support connectivity of VLAN networks between
multiple switches and to control plane nodes.  Currently only switch
ports corresponding to neutron ports are configured by the generic
switch driver.  This is a problem for some switches that do not support
adding a port to a VLAN that has not been created on the switch.

This change adds support for configuration of trunk ports when networks
are created and destroyed.  These trunk ports are configured via the
per-switch configuration option 'ngs_trunk_ports', which is a
comma-separated list.  Currently no drivers support this option.

Change-Id: I6c40f7f31058bc6e5e893743b6cdcd896d4850d1
Partial-Bug: #1690115