Detection of config errors for netmiko drivers
TODO: Do we need to support marking errors as retryable, and adding a
retry mechanism for this case?

NGS currently provides very little support for verifying that the
required configuration has been applied to a device. Commands are
formatted and sent to a device, but no validation is performed on the
response from the device. There are times when applying configuration
can fail, such as story 1737017, but this error is silently ignored.

This changes adds a simple error detection framework to the NGS
netmiko-based drivers. Command line error signatures are inherently
device-specific, but the framework introduced here is common.

A check_output method is added to the base Netmiko driver, which
compares the response from a device with a list of error message regular
expressions in self.ERROR_MSG_PATTERNS. Device drivers should populate
the patterns. On detection of an error, a
GenericSwitchNetmikoConfigError exception is raised.

Change-Id: Id7e3c1ed0c4f78d35c9662c9c52e12519905dfb5
Story: 2003148
Task: 23283
(cherry picked from commit d5f7ab31215dbae0d8ef262df502658922db2e9a)