Create and delete networks in parallel
Currently, networks are created and deleted on devices serially. In
systems with many devices, this can lead to these requests taking a long
time to complete.

This change executes device configuration in parallel for network
creation and deletion, which should improve this from O(N) to O(1). It
is not necessary to execute other operations in parallel, since they
only affect a single device. If we ever support multi-chassis LAG
configuration, we could configure it in parallel.

Failure handling should be as before this change, with the exception
that network creation will no longer stop after the first failed device,
since all are configured in parallel.

Change-Id: Ib257f4e7710ed6615e68c7034ff4e223bbf677b2
Story: 2006789
Task: 37328