WIP: Attempt to batch up cmds
When you have around 60 baremetal nodes attached to a single switch, it
takes a long time to execute all those commands. This gets worse when
you limit the number of concurrent ssh connections.

Here we look to batch up commands to send to the switch together, then
wait for them all to complete before returning the result. It does this
by using etcd keys as a queueing system. We pull off the queue using the
version at which the keys were added, giving a FIFO style queue.

Change-Id: I8c458bbc94df5630cfede5434bcdbe527988059c