Update Juniper driver
Recent release of JunOS have changed the command line syntax which
breaks the old driver. This change updates the driver to use the
new syntax. A further improvement would be to detect the JunOS version
and set the syntax accordingly.

This patch also adds some additional clean up to ports.