Add Dell Enterprise SONiC CLI driver
This patch is required urgently for Cambridge. It is unlikely
that we can wait for the upstream patch to merge:
https://review.opendev.org/c/openstack/networking-generic-switch/+/965373

This driver uses the `sonic-cli` rather than Linux userspace to
configure the switch. This is because LLDP advertises switch
ports based on the naming from the `sonic-cli`, which is not the
same as Linux userspace.

Note the existing Dell SONiC driver is using the Linux namespace
netmiko driver, which is not suitable.

Change-Id: Ia79de43a33c19f7b80bd0ef29badee464cd6fc0c
Signed-off-by: Doug Szumski <doug@stackhpc.com>