Add minimal Dell SONiC driver
A Dell SONiC driver has been added upstream:
https://review.opendev.org/c/openstack/networking-generic-switch/+/955252

This change adds a minimal Dell SONiC driver downstream. This is essentially
the existing SONiC driver, but with a slightly different enable mode command.

The change to the enable mode may not be required by the time we end up using
the upstream driver. This is because the upstream driver has been developed
against SONiC 4.5 which is more recent than the version used to develop this
patch.

This driver requires the switch to be configured to use the standard or
standard-extended naming mode.

Change-Id: Id263aa2757152012afdd6facb9f5fcd78fcc7ad8